Rename onnx_caffe2 to onnx
