+++ caffe2/python/onnx/helper.py
@@ -13,8 +13,8 @@
 # limitations under the License.
 ##############################################################################
 
-## @package onnx_caffe2
-# Module caffe2.python.onnx_caffe2.helper
+## @package onnx
+# Module caffe2.python.onnx.helper
 from __future__ import absolute_import
 from __future__ import division
 from __future__ import print_function
@@ -24,7 +24,7 @@ from caffe2.proto import caffe2_pb2
 from onnx import helper
 from onnx.backend.base import namedtupledict
 
-from caffe2.python.onnx_caffe2.workspace import Workspace
+from caffe2.python.onnx.workspace import Workspace
 
 import io
 import logging
