+++ caffe2/python/onnx/tests/optimize_onnx_test.py
@@ -13,8 +13,8 @@
 # limitations under the License.
 ##############################################################################
 
-## @package onnx_caffe2
-# Module caffe2.python.onnx_caffe2.tests.optimize_onnx_test
+## @package onnx
+# Module caffe2.python.onnx.tests.optimize_onnx_test
 
 from __future__ import absolute_import
 from __future__ import division
@@ -34,9 +34,9 @@ import numpy as np
 import onnx
 from onnx import helper, ModelProto, TensorProto
 from onnx.backend.test.runner import Runner
-import caffe2.python.onnx_caffe2.backend as c2
+import caffe2.python.onnx.backend as c2
 
-from caffe2.python.onnx_caffe2.tests.test_utils import TestCase
+from caffe2.python.onnx.tests.test_utils import TestCase
 
 class TestRoundtrip(TestCase):
     def _roundtrip(self, model_name):
