+++ caffe2/python/onnx/tests/test_utils.py
@@ -13,8 +13,8 @@
 # limitations under the License.
 ##############################################################################
 
-## @package onnx_caffe2
-# Module caffe2.python.onnx_caffe2.tests.test_utils
+## @package onnx
+# Module caffe2.python.onnx.tests.test_utils
 
 from __future__ import absolute_import
 from __future__ import division
@@ -24,7 +24,6 @@ from __future__ import unicode_literals
 import unittest
 
 import numpy as np
-import psutil
 
 
 class TestCase(unittest.TestCase):
@@ -35,11 +34,6 @@ class TestCase(unittest.TestCase):
         self.assertEqual(len(outputs1), len(outputs2))
         for o1, o2 in zip(outputs1, outputs2):
             np.testing.assert_almost_equal(o1, o2, decimal=decimal)
-
-    @staticmethod
-    def report_mem_usage(tag):
-        print('Mem usage ({}):'.format(tag),
-              psutil.virtual_memory())
 
     def add_test_case(name, test_func):
         if not name.startswith('test_'):
