--- caffe2/python/onnx_caffe2/tests/onnx_backend_test.py
@@ -1,49 +1,0 @@
-# Copyright (c) 2016-present, Facebook, Inc.
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-#     http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-##############################################################################
-
-## @package onnx_caffe2
-# Module caffe2.python.onnx_caffe2.tests.onnx_backend_test
-
-from __future__ import absolute_import
-from __future__ import division
-from __future__ import print_function
-from __future__ import unicode_literals
-
-import os
-
-import unittest
-import onnx.backend.test
-
-import caffe2.python.onnx_caffe2.backend as c2
-
-# This is a pytest magic variable to load extra plugins
-pytest_plugins = 'onnx.backend.test.report',
-
-backend_test = onnx.backend.test.BackendTest(c2, __name__)
-# Skip vgg to speed up CI, skip others because onnx-caffe2 does not support them yet.
-ci_blacklist = (r'(test_vgg19|test_vgg'  # Speed up CI.
-                '|test_ceil.*|test_floor.*'  # Does not support Ceil and Floor.
-                '|test_hardsigmoid.*|test_pow.*'  # Does not support Hardsigmoid and Pow.
-                '|test_mean.*|test_hardmax.*'  # Does not support Mean and Hardmax.
-                ')')
-if 'CI' in os.environ:
-    backend_test.exclude(ci_blacklist)
-# import all test cases at global scope to make them visible to python.unittest
-globals().update(backend_test
-                 .enable_report()
-                 .test_cases)
-
-if __name__ == '__main__':
-    unittest.main()
